Extract NostrDisplayNameService from ChatViewModel (-18 lines)
Continue phase 2.

WHAT WAS EXTRACTED
==================
Moved Nostr display name generation to NostrDisplayNameService (18 lines reduced):
- displayNameForNostrPubkey() (15 lines) → thin wrapper (6 lines)
- activeChannelDisplayName() (7 lines) → thin wrapper (3 lines)

NEW FILE
========
bitchat/Services/NostrDisplayNameService.swift (61 lines)

Features:
- Generates display names for Nostr participants
- Format: 'nickname#abcd' with collision-resistant suffix
- Detects self identity in geohash channels
- Uses known nicknames from geoNicknames map
- Fallback to 'anon#abcd'

API:
- NostrDisplayNameService.displayName(forPubkey:currentGeohash:currentNickname:geoNicknames:) -> String
- NostrDisplayNameService.channelDisplayName(_ channel:) -> String

IMPACT
======
ChatViewModel: 5,188 → 5,170 lines (-0.3% this commit, -16.5% cumulative)
New service: +61 lines
Tests: ✅ All 23 passing

Phase 2 progress: -178 lines
Remaining to < 5,000: ~170 lines
